refactor process_attestations
