+++ consensus/state_processing/src/common/is_attestation_same_slot.rs
@@ -1,11 +1,7 @@
+use safe_arith::SafeArith;
 use types::{AttestationData, BeaconState, BeaconStateError, EthSpec};
 
 /// Checks if the attestation was for the block proposed at the attestation slot.
-///
-/// Returns true if:
-/// - The attestation is for slot 0 (genesis), OR
-/// - The attestation's beacon_block_root matches the block actually proposed at that slot
-///   AND it's different from the previous slot's block (indicating no skip)
 pub fn is_attestation_same_slot<E: EthSpec>(
     state: &BeaconState<E>,
     data: &AttestationData,
@@ -14,8 +10,9 @@ pub fn is_attestation_same_slot<E: EthSpec>(
         return Ok(true);
     }
 
-    let is_matching_block_root = data.beacon_block_root == *state.get_block_root(data.slot)?;
-    let is_current_block_root = data.beacon_block_root != *state.get_block_root(data.slot - 1)?;
+    let is_matching_block_root = &data.beacon_block_root == state.get_block_root(data.slot)?;
+    let is_current_block_root =
+        &data.beacon_block_root != state.get_block_root(data.slot.safe_sub(1)?)?;
 
     Ok(is_matching_block_root && is_current_block_root)
 }
